Create DOM layout for lobby options
Currently does not actually perform any useful operations with the data.
Validation layer will be added.

The live component was changed to function components, because the
individual input components could not hook into phx-change (neither with
the Phoenix.HTML helpers, nor with barebones <input> tags) and writing
JS as a workaround seemed silly.